+++ packages/react/src/date-picker/docs/accessibility.mdx
@@ -1,0 +1,1 @@
+This component implements the accessibilty specifications mentioned in the [Date Picker Dialog Example](https://www.w3.org/WAI/ARIA/apg/patterns/dialog-modal/examples/datepicker-dialog/) by the [ARIA Authoring Practices Guide (APG)](https://www.w3.org/WAI/ARIA/apg/).
