+++ packages/react/src/date-range-picker/docs/code.mdx
@@ -1,0 +1,3 @@
+## Props
+
+<ComponentPropsTable name="DateRangePicker" />
