+++ packages/react/src/date-range-picker/docs/overview.mdx
@@ -1,0 +1,241 @@
+---
+title: Date range picker
+description: The Date range picker component allows users to select a range of dates via a calendar or text input.
+group: forms
+storybookPath: /story/forms-daterangepicker--basic
+figmaGalleryNodeId: 12444%3A100327
+relatedComponents: ['date-picker']
+---
+
+<DoHeading />
+
+- always display the date format above the text input
+- use Australian date format 'dd/mm/yyyy'
+- allow users to navigate dates via the calendar
+- ensure users can enter dates via the calendar or text input
+- Make sure error messages are specific about what input is required.
+
+<DontHeading />
+
+- never hide the date format
+- use international date formats such as USA mm/dd/yyyy
+- hijack built-in keyboard navigation behaviour.
+
+## Date range picker
+
+`DateRangePicker` is a [controlled component](https://reactjs.org/docs/forms.html#controlled-components) which means consumers of this component need to manage the state of this component by using the `value`, `onChange`, `onFromInputChange` and `onToInputChange` props.
+
+For an example of using this component in a form built with `react-hook-form` and `yup`, please see the [Single-page form template](/templates/single-page-form).
+
+```jsx live
+() => {
+	const [value, setValue] = React.useState({ from: undefined, to: undefined });
+	return (
+		<DateRangePicker
+			value={value}
+			onChange={setValue}
+			onFromInputChange={(from) => setValue({ ...value, from })}
+			onToInputChange={(to) => setValue({ ...value, to })}
+		/>
+	);
+};
+```
+
+## Tracking the input value
+
+The `DateRangePicker` component has two methods of input:
+
+1. Typing in a date string (dd/mm/yyyy) via the text inputs
+2. Selecting a date via the calendar widgets
+
+As we do not have any sort of input masking due to accessibility concerns, there is a possibilty for the user to enter invalids value via the text inputs.
+
+You can use the `onFromInputChange` and `onToInputChange` props to keep track of the users input. The `value.from` and `value.to` props can also be set to a string, which represents the value of the text inputs.
+
+```jsx live
+() => {
+	// Set the value to a value that the user might think is valid
+	const [value, setValue] = React.useState({
+		from: '30/1o/2020',
+		to: '31/1o/2020',
+	});
+
+	// This logic is for documentation purposes only. This should be done with `yup` or `zod`.
+	// See Single-page form template
+	const isInvalid = React.useCallback((value) => {
+		if (typeof value === 'undefined' || value == '') return false;
+		if (value instanceof Date && !isNaN(value.getTime())) return false;
+		return true;
+	}, []);
+
+	const fromInvalid = isInvalid(value.from);
+	const toInvalid = isInvalid(value.from);
+
+	return (
+		<DateRangePicker
+			value={value}
+			onChange={setValue}
+			onFromInputChange={(from) => setValue({ ...value, from })}
+			onToInputChange={(to) => setValue({ ...value, to })}
+			fromInvalid={fromInvalid}
+			toInvalid={toInvalid}
+			message={fromInvalid || toInvalid ? 'Enter a valid date' : undefined}
+		/>
+	);
+};
+```
+
+## Legend
+
+Use the `legend` prop to describe the purpose of the group of fields.
+
+```jsx live
+() => {
+	const [value, setValue] = React.useState({ from: undefined, to: undefined });
+	return (
+		<DateRangePicker
+			value={value}
+			onChange={setValue}
+			onFromInputChange={(from) => setValue({ ...value, from })}
+			onToInputChange={(to) => setValue({ ...value, to })}
+			legend="Date period"
+		/>
+	);
+};
+```
+
+## Hint
+
+Use the `hint` prop to provide help that’s relevant to the majority of users, like how their information will be used, or where to find it.
+
+Don't use long paragraphs and lists in hint text. Screen readers read out the entire text when users interact with the form element. This could frustrate users if the text is long.
+
+Don't include links within hint text. While screen readers will read out the link text when describing the field, they will not tell users that the text is a link.
+
+```jsx live
+() => {
+	const [value, setValue] = React.useState({ from: undefined, to: undefined });
+	return (
+		<DateRangePicker
+			value={value}
+			onChange={setValue}
+			onFromInputChange={(from) => setValue({ ...value, from })}
+			onToInputChange={(to) => setValue({ ...value, to })}
+			legend="Date period"
+			hint="Example hint text"
+		/>
+	);
+};
+```
+
+## Invalid
+
+Use the `fromInvalid` and `toInvalid` props to indicate if the user input is invalid (does not validate according to the elements settings).
+
+```jsx live
+() => {
+	const [value, setValue] = React.useState({ from: undefined, to: undefined });
+	return (
+		<FormStack>
+			<DateRangePicker
+				value={value}
+				onChange={setValue}
+				onFromInputChange={(from) => setValue({ ...value, from })}
+				onToInputChange={(to) => setValue({ ...value, to })}
+				fromInvalid={true}
+				toInvalid={true}
+				legend="Date period"
+				message="Enter a valid date"
+			/>
+			<DateRangePicker
+				value={value}
+				onChange={setValue}
+				onFromInputChange={(from) => setValue({ ...value, from })}
+				onToInputChange={(to) => setValue({ ...value, to })}
+				fromInvalid={true}
+				toInvalid={false}
+				legend="Date period"
+				message="Enter a valid date"
+			/>
+			<DateRangePicker
+				value={value}
+				onChange={setValue}
+				onFromInputChange={(from) => setValue({ ...value, from })}
+				onToInputChange={(to) => setValue({ ...value, to })}
+				fromInvalid={false}
+				toInvalid={true}
+				legend="Date period"
+				message="Enter a valid date"
+			/>
+		</FormStack>
+	);
+};
+```
+
+## Custom labels
+
+Use the `fromLabel` and `toLabel` props to change the field labels.
+
+```jsx live
+() => {
+	const [value, setValue] = React.useState({ from: undefined, to: undefined });
+	return (
+		<DateRangePicker
+			value={value}
+			onChange={setValue}
+			onFromInputChange={(from) => setValue({ ...value, from })}
+			onToInputChange={(to) => setValue({ ...value, to })}
+			fromLabel="From"
+			toLabel="To"
+		/>
+	);
+};
+```
+
+## Disabled
+
+Disabled input elements are unusable and can not be clicked. This prevents a user from interacting with the input element until another action is complete.
+
+```jsx live
+() => {
+	const [value, setValue] = React.useState({ from: undefined, to: undefined });
+	return (
+		<DateRangePicker
+			value={value}
+			onChange={setValue}
+			onFromInputChange={(from) => setValue({ ...value, from })}
+			onToInputChange={(to) => setValue({ ...value, to })}
+			disabled
+		/>
+	);
+};
+```
+
+## Minimum and maximum dates
+
+The `minDate` property can be used to disable any days before a specific date.
+
+The `maxDate` property can be used to disable any days after a specific date.
+
+If a valid date is entered using the text input but it falls outside the constrained range, the closest valid date will be used.
+
+```jsx live
+() => {
+	const [value, setValue] = React.useState({ from: undefined, to: undefined });
+
+	const today = new Date();
+	const lastWeek = new Date(today.getTime() - 7 * 24 * 60 * 60 * 1000);
+	const nextWeek = new Date(today.getTime() + 7 * 24 * 60 * 60 * 1000);
+
+	return (
+		<DateRangePicker
+			value={value}
+			onChange={setValue}
+			onFromInputChange={(from) => setValue({ ...value, from })}
+			onToInputChange={(to) => setValue({ ...value, to })}
+			minDate={lastWeek}
+			maxDate={nextWeek}
+		/>
+	);
+};
+```
