+++ packages/react/src/date-range-picker/index.ts
@@ -1,0 +1,1 @@
+export * from './DateRangePicker';
