+++ packages/react/src/indicator-dot/IndicatorDot.test.tsx
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom';
 import 'html-validate/jest';
 import { cleanup, render } from '../../../../test-utils';
-import { IndicatorDot, IndicatorDotProps } from './IndicatorDot';
+import { IndicatorDot, IndicatorDotProps } from '../badge';
 
 afterEach(cleanup);
